+++ tools/azsdk-cli/Azure.Sdk.Tools.Cli/CHANGELOG.md
@@ -1,5 +1,34 @@
 # Release History
 
+## 0.5.2 (Unreleased)
+
+### Features
+
+- Added new tool to update language exclusion justification and also to mark as language as excluded from release.
+
+### Breaking Changes
+
+- None
+
+### Bugs Fixed
+
+- None
+
+## 0.5.1 (2025-10-07)
+
+### Features
+
+- None
+
+### Breaking Changes
+
+- None
+
+### Bugs Fixed
+
+- Create release plan tool failure
+- Use existing release plan instead of failing when a release plan exists for spec PR.
+
 ## 0.5.0 (2025-09-25)
 
 ### Features